Fix the detail screen + minor cleanup to section-level properties
This commit includes **multiple minor fixes to the detail screen**:
- Fixes `getLocalTimeString` so that it doesn't modify the incoming datetime.
  Such modification would be an unintended side effect. This fixes the "Invalid
  date" issue on the detail screen.
- Stops adding incidents to the trip in the detail screen. This fixes the
  "`digest` called 10 times error related to the geojson for the map"

Next, it **moves `getColor` from the list to the service**. This allows us to
call `getColor` from the detail view as well, and ensure that the icons in the
detail view area also colored.

Finally, it involves some **refactoring** to remove function calls from the HTML.
These functions are called in the controller instead, and the resulting
properties are displayed in the HTML. This should improve performance since we
don't recalculate values every time the user interacts with the UI.

Most of these changes were in the detail HTML, since I already made similar
change to the list view in 5eae45953ee1cc0a5c0051c700fb6b6717c96021. But we
were still calling `setColor` to set the icon color. Fixed that as well by
refactoring `getPercentages`

- Also removed duplicate locations that mapped `mode` -> `icon` and `mode` -> `color`,
unifying them into `getIcon` and `getColor` methods that could be invoked
elsewhere. Note that this also unified minor differences between the locations
(e.g. for TRAM, `slateblue` for the line and `darkslateblue` for the icon
colors in `list.js`) caused by copy-pasted code.

- Also removed unused function `allModes`