Unify mode and purpose code
The mode and purpose code had a lot of copy-pasted components that were
different only in variable names. Copy-pasting is bad and does not follow the
DRY principle.

And this is only going to get worse as we want to add counter-factual mode
selection as a third dropdown.

Prior to that, we refactor the confirmation code to be DRY.
The list of confirmation keys is defined in `www/js/tripconfirm/trip-confirm-services.js`
and the corresponding options are in `www/json/trip_confirm_options.json.sample`

At this point, we also have separate popover templates in `www/templates/diary`
(in case we want the text to be different) and separate buttons in
`www/templates/diary/list.html`

Everything else takes the `inputType` as a parameter or iterates over all input
types and is completely generic.

Testing done:
- on both android and iOS:
    - created trips to show in the diary
    - selected various pre-defined mode and purpose and "other" options
    - pre-defined modes were selected
    - other options were selected and added to the list
    - cancelling the other popup also worked :)