Bump up version numbers again due to stupid app store connect bug
I tried to submit for app store review, but I was not able to see the build after uploading.
After uploading 5 separate builds, and downgrading my version of xcode, I
finally figured out that the issue was that I had started preparing the
release while the archive was being created.

https://stackoverflow.com/a/63951622/4040267

I could not find a way to delete the partially created release, so I changed
the version in xcode to v1.0.2 and retried. And it worked.

I then changed the app store release version also to v1.0.2, associated the
build and submitted for review.

Bumping up the version numbers to reflect this.