Store a bunch of device settings that can help debug any tracking issues on the phone
I was going to store the language in the profile as part of the CEO ebike
project, so that we can serve separate English and Spanish surveys. And since I
was adding a background updater to the startup anyway, I though that I might as
well add other device settings that may help debug tracking issues, such as the
app version, the OS version, and the manufacturer (for https://dontkillmyapp.com/).

This fixes an issue requested by @asiripanich as part of the UNSW study, but I
am not able to find it to auto-close :(

Testing done:
- Integrated it into the CanBikeCo branch
- Ensured that the values were saved into the profile database
- Thanks to the magic of autogenerated tables, the new fields even show up in
  the participant view

- Dropped the connection
- Launched the app again
- Ensured that an error message was generated