Fix regression in enhanced trip matching
The enhanced trip matching (75129dbbe4e7d85e7e263f550ec96b3e078c6c01) caused a
regression in which a spurious trip (not a trip) that occurred after the real
trip fit the criteria for a match.

And since it was confirmed after the real trip, as you would expect while going
down the trip list, it was matched preferentially.
https://github.com/e-mission/e-mission-docs/issues/476#issuecomment-747587041

Fixed by checking the degree over overlap and rejecting too short matches