Display user inputs (almost) correctly
Display user inputs correctly in the situation that they are in the confirmed
trip objects

- call `populateInputFromTimeline` for all trips
- find the `userInputLabel` directly from the trip
- set up everything else accordingly, notably the structure while creating the popup

However, this doesn't work in the case where the user has set the labels and
the pipeline has not yet run. In that case, if the user refreshes the page, all
the currently set labels will "disapper".