Support filtering
- Replace the date button with two buttons for unlabeled and invalid trips
- Add filter functions
- Implement select/deselect logic

NOTE that for a large number of trips, this is very slow due to the log
statements in the user matching.

We should address performance issues in the next step.
- Improve performance of the map display by returning only a sample of data
- Improve performance of the user label matching by only retrieving potential matches

Both of these imply that we need to support querying by data.ts and not just
metadata.write_ts, which is a server-side change.