Always match against the unprocessed labels
We want to always match against the unprocessed labels, even when the confirmed
trip has labels, because the unprocessed labels might be newer.

For example, consider the case in which the user corrects a previously confirmed
trip. Until the data is pushed and the pipeline runs again, the confirmed trip
will have the old value, which is incorrect.

Fortunately, we no longer have to retrieve all the user labels, only ones after
the last pipeline run. If the pipeline has run, the confirmed trip should have
been updated with the info.

Since the pipeline currently runs every hour, the number of new inputs
likely to be small, leading to:
- less redundant information
- fewer labels to check
and an overall more scalable solution

Testing done:
- Changed a label to bikeshare
- Reloaded

Without the fix, the entry reset to "Walk"
With the fix, the entry is "Bikeshare"

- Ran the pipeline
- Reloaded

Still saw a user input. WTF!
But realized that it will be gone after the next real trip
since we use the confirmed_trip timestamp as the last pipeline_ts