Improve the performance of the trip detail screen
In conjunction with some simple server-side changes, this allows us to retrieve
a small subset of the location information. This improves the performance
dramatically.

Without this fix:
on expanding an hour long trip, we read 3090 entries, and took close to a second, even on a local server
(`0.8326890468597412` secs)

With the fix:
we read 442 entries in `0.2554280757904053` secs

Lowering down to 100 points,
we read 51 entries in `0.028665781021118164` secs

No hugely visible difference in the visualization (see the PR)

Using 100 points as the default and turning zoom off since we don't have the
ability to maintain quality at fine grained zoom.

The server-side fix also supports retrieving by `date.ts` instead of
`metadata.write_ts`, which allows us to retrieve the recreated locations
instead of the filtered locations, and avoid bad gaps in the trip visualization