Handle gaps in trips
Without this change, if there is more than a one week gap in the middle of a
timeline, the infinite scroll stops. This is because we are not retrieving a
certain number of trips in each batch, but we assume that there are no more
trips if our batch gets zero entries.

So, we use new server-side changes to retrieve both the last element and the
first element on the server. We only end the scroll when the first trip that we
see is the first trip that the pipeline processed.

+ new wrapper to call the new server method
+ minor changes to return values
+ some additional control to update the endTs in the zero-batch-but-no-trip-end
case. We just bump up the currentEndTs by a week.