Minor fix with big but corner case impact
Now that we rely on the infinite scroll to load more entries, we need to ensure
that the displayList is also zeroed out on setup. Otherwise it looks like there
are plenty of entries, and the infinite scroll won't kick in.

So if you log out and login, then the trips don't get replaced.

This is definitely a corner case, since most users are not going to login to
somebody else's account. But if people want to refresh, they need to be able to
see the newly refreshed value not whatever they had before.