Convert the diary from auto-refresh to manual refresh
The manual refresh was cool when we started, but it had a lot of problems after
we added the popups. Using the popup on the first entry would frequently
trigger the refresh, which would close the popup. And launching the popup again
would trigger the refresh again, ad-infinitum.

Instead, we use a manual refresh which is less cool but works well.