initial pass at refactoring the consent code (missing i18n)
- add a new button to the profile screen
- pull out the consent text into a separate file
- `ng-include` the consent text into the existing consent file
- create a new popover template and `ng-include` the consent text there as well
- convert the template load in intro.html back to a standard file for now

Next step: enable i18n

This will require us to move the `geti18nFile` method out of intro.js. I tried
a shortcut with `$rootScope`, but it doesn't work because `intro.js` is not
loaded if it is not the first initialization.

We might want to split up at least some parts of intro anyway as we work on the
onboarding and the status screen, since we will also want to access the status
screen from the onboarding and from the profile later.