Add a secret check before proceeding
This is a hack to support the specific requirements of the NREL publication
process, which is currently intended to be a closed system.
- Users need to scan a specified QR code that represents a secret
- They are allowed to proceed only if the secret matches the app secret

In the long term, we want to use channels directly instead of this "secret"
code. However, we didn't want to deal with externally hosted software for
approval at this time.

This is the commit fixes the first three tasks in
https://github.com/e-mission/e-mission-docs/issues/628

in particular,
https://github.com/e-mission/e-mission-docs/issues/628#issuecomment-799793421
and
https://github.com/e-mission/e-mission-docs/issues/628#issuecomment-799828333

Changes:
- new "secretcheck" service
- on receiving a custom URL, pass the parameter to secretcheck
- secretcheck validates the secret and stores it
- change startprefs so that if we have a valid stored secret, we can move on to
  checking the other onboarding states
- copy over the templates and the javascript code for the first screen from the
  em-base repository