Switch the infinite scroll from auto-load on scroll to a button
The auto-load on scroll feature, so common to messaging and chat apps is
actually fairly complicated to implement correctly in the case in which
we display a filtered list. In particular:
- the additional scrolling required to maintain "place" in the list triggers
  multiple "scroll" calls even when the user is not doing anything, which leads
  to multiple calls to the server.
- if new entries are not added to the list after the server call, then the
  condition that triggered the server call (top < threshold) persists, leading
  to multiple calls again.

See https://github.com/e-mission/e-mission-docs/issues/662

This is an inherent limitation of using auto-load on scroll for a
filtered list and also happens with the standard ionic component,
https://github.com/e-mission/e-mission-docs/issues/662#issuecomment-899312254

So instead of butting my head against the wall getting auto-load to work, I
switched to manual load, similar to the diary screen.
https://github.com/e-mission/e-mission-phone/pull/769

As a bonus, this also helps with https://github.com/e-mission/e-mission-docs/issues/659
It indicates how many of the trips are displayed in **this filter**, should
help to clarify questions around "missing trips" and allow users to explore
other tabs as necessary.

Testing done:
- loaded data for a user with a single trip
- labeled that trip so the user had no trips
- loaded data for a user with multiple trips

In all cases, used the button to scroll all the way to the oldest trip.
Did not encounter any issues.