Try to fix the walkthrough again
Albeit in a lamer fashion.
The problem is that there are multiple elements with the same ID in the list.
We could consider switching to a more sophisticated walkthrough solution.
However, for now, we just step out to the full list of trips instead of
highlighting individual items.

This seems to work pretty reliably, let's push it to staging and see if it
works.