Final cleanup of the main list code
Remove all vestiges of the ConfirmHelper code from `list.js`

```
$ grep ConfirmHelper www/js/diary/list.js  | wc -l
       0
```

Also remove the list of supported inputs from the directive to avoid
duplication with the list in the `ConfirmHelper`. Later, we should clean up
`ConfirmHelper` to read that from the config instead of putting it into code.

Or if we choose to pass it into the directive, we have to figure out how to
pass it in to the factory at creation time.

There is no point in configuring it in two places.