Pull the populate code out into a service
Pull out all the populateXXX and recalculateXXX functionality into a separate
service so it can be invoked from multiple controllers.
This is consistent with https://github.com/e-mission/e-mission-docs/issues/674#issuecomment-933083710

We load the service dynamically to reduce the hardcoded dependency issue.

We retain a couple of simple wrappers e.g. `fillUserInputsGeojson` in the
controller so that we can invoke it from `$watch` more easily and we can call
`$scope.$apply` to update the visualization.

Remove references to `scope.manualResultMap` from the controller since the
expanded values will be pre-populated.

Pass the viewScope around as needed to support the callback after the update.
Add a new NOP callback to the diary view to match the desired external interface.
Add blank inferred label datastructures for the diary pending a bigger
restructuring that will send inferred labels to as part of the geojson.