Simplify the invocation of `getUserInputForTrip`
Since `getUserInputForTrip` is now called only from `populateManualInput`
we can adapt it to take the trip inputs directly, and we don't have to massage
the input to meet expecations

The one minor ordering change to accomplish this was to compute the basic
classes before matching the inputs so that we can use the `isDraft` variable
directly. We don't (currently) show draft trips on the label screen, so it will
be undefined there. And undefined != true, so it will only apply for diary trips.