Fix the `verifyTrip` code to use `$scope.trip` instead of `trip`
Since the variable is no longer passed in.
And it can't be passed in because the `verifyTrip` is actually invoked from
outside this scope by the `verifycheck` directive.

We couldn't test this before because all the trips were already with green labels.
Tested with a yellow label and it worked.