[diary only] Support both imperial and metric units with a config option
**Note that this currently works only for the diary**. The dashboard has not
been fixed because it is not currently visible in the CEO branch :)

My original plan was to only support metric units, since

https://en.wikipedia.org/wiki/Metrication
"According to the U.S. Central Intelligence Agency's online The World Factbook
(2016), the only countries that have not adopted the metric system are Myanmar
(also known as Burma), Liberia and the United States."

However, it looks like this project will be US-focused for the forseeable
future, and even the UK is beginning to revert its metrication process. So we will support both imperial and metric units with a config option.

- Create a new service for this conversion
- Have it use metric or imperial based on a config flag
- Change all existing distance calculations to use the new service
- Add a "suffix" to the service to distinguish between "mi" (miles) and "km"
  (kilometers)
- Change all the templates and the i18n to use the suffix instead of hardcoded values

Testing done:
- set the config value to false; confirmed that everything is in km
- set the config value to true; confirmed that everything is in miles and there
  were no errors in the console log