Allow users to resize the items
This fixes issues in which the labels are partially hidden on some phones, and
is a reasonable workaround for the inability to test with multiple phones.
We could even consider allowing users to save their perfect height in a local
setting later so they don't have to keep changing the settings every time.

I tried to do this automatically, but failed with the current UI
https://github.com/e-mission/e-mission-docs/issues/679#issuecomment-944420404