Turn off the display of the automated mode detection
Based on feedback from the Smart Commute program that the automated modes were
confusing participants. We may want to reintroduce them later once we start
including automated results into the metrics.

This is fairly hacky, and if it is a frequent request, we may want to add it to
our future config file.