Add the new unused app restrictions to the status screen
Fairly straightforward change similar to prior changes for
location and motion activity.

The one "special" difference is that we move the buttons out of the directive
and into the top level intro, which is where it makes sense in the long term
anyway.

We should have separate top level directives for the individual sensors that we
use.