UI changes to handle the "deny" use case correctly
This is the complement to
https://github.com/e-mission/e-mission-data-collection/pull/195/commits/61a7b8b99ffecd22acc41e9644429b75f5f8312f

It primarily changes the instructions displayed to the user to be consistent
with the actual operations they need to do.

TODO: Do something more sophisticated than just displaying the error message.
Return an error code in addition to the message and use it to look up the UI displayed message.