Extract mode labels for user inputs
From
https://github.com/e-mission/e-mission-server/pull/844/commits/ab8211600bc1c10843a335aa1bc96f960f76ac95,
the modes from user inputs are now included in the metrics with a `label_` prefix.

So we change the metrics mapping code to handle these inputs and return the
appropriate map. Note that we can also use this to determine which CO2 mode to
use, but we haven't done this yet.