Fix the summary functions and precompute summary values as well
`getSummaryData` and `formatData` really need to be unified, but it's a
bit tricky since `getSummaryData` assumes a single numeric value while
`formatData` assumes a list of lists. And `getSummaryData` replaces the
single numeric value with a single string value.

Let's get the footprint and calorie counts to work at a basic level and then
come back to the refactoring.

For now, we just precompute summary values for user and aggregate metrics, and
add per capita aggregate metric calulations.

Follow-on to 476206dc74e0c96328a2ba7c4f850bd7acc591b8