Pass in the mode maps or summaries to the calorie and carbon footprint calculations
So that we don't have recompute summaries and/or re-map every time. This should
improve client performance.

This also allows us to remove `getAvgSummaryDataRaw`, which is a copy-paste of
`getSummaryDataRaw` that was created solely for the purpose of getting per
capita aggregate footprints (e.g. call `getDataFromMetrics` with
`metric2valAvg` instead of `metric2val`). The rest of code was identical.

We can now pass in the per capita values to `getSummary` instead, which
obviates the need for this function.

Follow-on to 6850bd49402c2a4d3cf506bb35aea95d5b973832