Refactor `formatData` to a set of formatters that can be reused
We need to format both the chart data and the summary data, but the summary
card wants one value per mode while the charts want multiple values per mode.
So the format code needs to work on an array in one case and on individual
values in the other.

So we pull out the formatters such that they work on one value at a time, which
will also simplify the giant bolus of code that is the
formatXXX/getSummaryData. We can then call the formatters as the inner value in
the loop, or maybe even call it directly from the HTML

https://github.com/e-mission/e-mission-docs/issues/688#issuecomment-1000356833