Call the formatters directly from the HTML for the numeric summary tab
Now that we have specific formatters, we can just call them directly from the
HTML and don't have to pre-compute. And obviously, the easiest way to do this
is to format the precomputed summaries.

So it turns out that we do implement the second option from
https://github.com/e-mission/e-mission-docs/issues/688#issuecomment-1000356833
after all, just directly from the HTML!!