Partially implement the tiered proposal with one "range-limited motorized vehicle"
Additional description:
https://github.com/e-mission/e-mission-docs/issues/688#issuecomment-1000595640

- while reading the footprint code, identify which entry is the range-limited
motorized vehicle (RLMV)
- ensure that we have only one of these
- handle the various cases (default, custom without RLMV, custom with RLMV) and
  return the ranges and the optimal footprint for each range

Note that this doesn't actually work
https://github.com/e-mission/e-mission-docs/issues/688#issuecomment-1000626564
so it is not invoked from anywhere.

Instead, we just have the code in place in case we want to enable it in the
future, or, more likely, to port it to the server.

Testing done:

**Checked logs**

```
Found range_limited_motorized mode {text: "Pilot e-bike",
    value: "pilot_ebike", met: {…}, co2PerMeter: 0.00728, range_limit_km: 30, …}
    $$hashKey: "object:612"co2PerMeter: 0.00728met: {ALL: {…}}range_limit_km: 30
    text: "Pilot e-bike"value: "pilot_ebike"__proto__: Object
```

**At a breakpoint, tried to invoke the function manually**

```
FootprintHelper.getOptimalDistanceRanges()
Non motorized mode or footprint <= range_limited_motorized walk 0 0.00728
Non motorized mode or footprint <= range_limited_motorized pilot_ebike 0.00728 0.00728
Non motorized mode or footprint <= range_limited_motorized bike 0 0.00728
Non motorized mode or footprint <= range_limited_motorized bikeshare 0 0.00728
Non motorized mode or footprint <= range_limited_motorized not_a_trip 0 0.00728
Non motorized mode or footprint <= range_limited_motorized other 0 0.00728
0: {low: 0, high: 5000, optimal: 0}
1: {low: 5000, high: 30000, optimal: 0.00728}
2: {low: 30000, high: 600000, optimal: 0.00894}
3: {low: 600000, high: 1.7976931348623157e+308, optimal: 0.1}
length: 4
__proto__: Array(0)
```

Note that we do return 4 ranges, but the optimal footprint for the 5k to 60k
range is `scootershare`, which is not really a long-term mode. This indicates
another problem with this approach.

Fixing it would require additional metadata for each mode (is it long-range or
range-limited, etc), which is beyond the scope of this change.
https://github.com/e-mission/e-mission-docs/issues/688#issuecomment-1000664738