Display a range of values for the carbon footprint
If there are modes for which we don't have a mapping, including "unlabeled" or
any of the custom modes, we don't just ignore it. We generate a range from
where the footprint is 0 to the footprint is the max of the available modes.

https://github.com/e-mission/e-mission-docs/issues/688#issuecomment-1000667060
https://github.com/e-mission/e-mission-docs/issues/688#issuecomment-1000682256
https://github.com/e-mission/e-mission-docs/issues/688#issuecomment-1000721270

Includes:
- new custom directive to display the difference properly, which took a hour to
  get the spacing right
- new custom directive to display the range properly

In both cases, we need to ensure
    - if the two values are the same, we only display one value
    - if the two values are different, we display both values

For the difference code, we additionally need to address the sign of the change in the text
https://github.com/e-mission/e-mission-docs/issues/688#issuecomment-1000692615

Also includes:
- new text to display the range, and specially to get the spacing of the range correct
I hate spacing.