Change the calorie calculations to also have a range
Very similar to 42e8173f329bcfc86bfc991425a6c137c9cf45cf and uses some of the
similar themes

- if there are modes for which we don't have an MET, including "unlabeled" or
  any of the custom modes, we don't just ignore it. We generate a range from
  where the MET is 0 to the MET is the max of the available modes

Uses the new custom directives defined in
42e8173f329bcfc86bfc991425a6c137c9cf45cf to display the differences and ranges
properly

Change the text for the food comparisons, to say "at least" and use the lower
value. Changing the text to support the full range would be more complex since
if the values were the same, we would have to fall back to the 1 banana/2
bananas change, but if the low and high were different, we could just use
"bananas" throughout. I don't know how to do that in the translate logs, and I
don't want to deal with the ng-if statements to deal with it in the HTML