Handle the case where no trips have been labeled for two weeks
This is likely fairly common for those people who have given up on labeling, so
it is worth addressing. The problem is that in that case,

```
lastWeek.footprint.low = 0
thisWeek.footprint.low = 0

thisWeek.footprint.low/lastWeek.footprint.low = NaN
```

So we end up with a blank "low" on the diff display, which looks weird.

To fix, we add a pre-processing step to the directive, which converts NaN -> 0
and display the % if it is >= 0

This then leads to the marginally less weird case in which we don't show the
arrow and we don't show the OR

To fix that as well, we remove the `ng-if` for the OR (not sure why it was ever added)
and add an = for the case in which the metric is 0