Internationalize the sensor headings as well
For most of these, we can translate the values directly since they are not
platform specific. The fitness name is platform specific so we compute it in
the controller, just like all the other names.

This addresses beta testing feedback.

"The wording of the movement permission and the unused apps permission were a
bit different on my screen, and it took me a bit to figure out the unused apps
permission. Perhaps the heading and instructions for those two permission items
could be worded more closely to how they're each labeled in my two screenshots
attached."