Refresh the status checks before the status modal is opened
Before this change, the status was only computed when the controller was
initialized. This is fine for all permission related changes, since downgrading
permissions causes the app to restart and the controller to get re-initialized.
However, changing location settings does not cause the app to restart, and
presumably other status checks might not as well.

So if we initialize the controller (maybe with green values) and then fail the
status again, then it stays green.

The principled fix for this is probably a service, which will also allow us to
read the status and change the color and icon accordingly.
https://github.com/e-mission/e-mission-docs/issues/680#issuecomment-1025179763

But for now, a decoupled approach, with event-driven message passing is be a
better option. We can revisit when we refactor the app status code into
individual components.

If we choose to keep an event-driven architecture, we can also use events to
get the overall status back to the parent component.