Ensure that the phone app headers are large enough
If they are too narrow, they cause the actual footprint to be hidden.
There is no known reason why they are an absolute, instead of a relative value.
They are only used in this one place, so changing them should not break anything else.
So we just change them back to a relative value

More details:
https://github.com/e-mission/e-mission-phone/pull/812#issuecomment-1048328735
https://github.com/e-mission/e-mission-phone/pull/812#issuecomment-1048359270
https://github.com/e-mission/e-mission-phone/pull/812#issuecomment-1048365598
https://github.com/e-mission/e-mission-phone/pull/812#issuecomment-1048461640