Only display the worst carbon and the aggregate carbon once they exist
As part of the humanize change d9c340e9ad045886db64f2476edb8ba9a5190cdd
we reformat the values in the scope. This means that when the input values are
changed, we can't detect that the `range-display` values have also been
recomputed.

So for elements that are created before the values are retrieved from the
server, the range low and high are blank.

We could `$watch` or something for the values to change and then call `$apply`
again, but an easier option is to just defer the initialization of the metrics
until the value is actually set.