Fix "second level of end checks" for the user label matching
As part of https://github.com/e-mission/e-mission-phone/commit/3ec573446b2e56f865750d3da03744c2aafed158
we changed the input from a trip geojson to a trip object. But we forgot to
change one of the code paths, which caused a regression.

Fixed by changing the variable and the format of the variable.

Testing done:
- loaded fake manual entries for a test user
- before the change, got the error
- after the change, did not get the error

This fixes https://github.com/e-mission/e-mission-docs/issues/713