Upgrade leaflet to the most recent 1.8.0
Since we are fixing maps anyway, there's no need to stick with an obsolete
version. The change is fairly straightforward - I don't think we were using a
lot of the complex functionality in leaflet

This is also installing a directly downloaded version instead of downloading
from a package manager, so should be smaller.

The only regression seems to be in events generated after the map has been
unloaded, in which case it doesn't have any bounds and our `console.log` fails.

We catch the `ValueError` error generated in that case and ignore it.

```
in mapEvents, event = "leafletDirectiveMap.detail.unload" leafletEvent = "unload" leafletObject is undefined
in mapEvents, event = "leafletDirectiveMap.detail.zoomlevelschange" leafletEvent = "zoomlevelschange" leafletObject is undefined
in mapEvents, event = "leafletDirectiveMap.detail.layerremove" leafletEvent = "layerremove" leafletObject is undefined
```