Support the "notch" properly on iOS
This fixes the second issue in
https://github.com/e-mission/e-mission-docs/issues/722#issue-1216238583

In particular, we add the `viewport-fit=cover` tag.
Unfortunately, since we are using ionic version 1, it doesn't automatically set
all the safe areas correctly.

So we had to figure out how to send them manually.
Finally got it to work, applied it only to iOS
https://github.com/e-mission/e-mission-docs/issues/722#issuecomment-1110322266

Both iOS and android work well now, at least in profile mode
https://github.com/e-mission/e-mission-docs/issues/722#issuecomment-1110343897