Add the labels to the detail screen
So you don't have to see the trip and then go back to the list to verify
Fortunately, this was already a directive, so the change largely involved
adding the directive to the detail screen.

Couple of modifications:
- we allow the specification of a linkedid to make the DOM traversal simpler
- we choose the survey opt in the detail screen as well
- on `recomputeDisplayTrips`, we close the detail screen

TODO: Might want to configure the wait to recompute since 30 secs is too large
for the detail screen use case.

https://github.com/e-mission/e-mission-docs/issues/111