Ensure that labels set in the diary screen are reflected in the label screen and vice versa
Since the diary and label screens are not currently unified, values set in one
screen are not reflected in the other. This is confusing if people switch
between the two.

Hacking around this by copying the user inputs from the diary to the list and
vice versa after enter.

Changes:
- new function to copy input if newer
- add a write_ts to modified user inputs to track which is newer
- call the new function for all trips in the diary and for trips that are
  loaded in the label (since the label has more trips than the diary).